iPCR-single-QA.bash; small bugfix
Previously I undid 'discard-untrimmed' in iPCR-trim.bash. This led to
ommision of some statistic in the stats file. The iPCR-single-QA script
assumed this statistic to be present and therefor crashed. I now removed
this statistic from this QA script as well.
Better would be to improve the iPCR-trim script to include the
discard-trimmed option again but now without exessive memory
requirements.