iPCR-bam2bedpe.bash; undone RevComp
Previously the bedpe files reported the original read sequences, also if
they were mapped to the minus strand. In the latter case the bamfile has
the reverse complement of the read sequence because SAM always reports
sequences relative to the forw strand. And thus, in this case I computed
the RevComp of those BAM-sequences to get back the original sequence.
But to find SNP I do need to work with sequences relative to the forward
strand, as SNPs are defined on that strand. So I commented out this code
making the reverse complement.
Now all sequences in the bedpe are reported relative to the forw strand.