SuRE-snakemake, iPCR-map.bash: specify statsfile
The WASP rules use different subdirectories. Therefor the stats written
by the script 'iPCR-map.bash' are (in the old script) also written to a
stats file in that sub directory. Now, the 'iPCR-map.bash' script takes
an extra argument (-t) specifying the full name of the stats file. The
SuRE-snakemake script specifies this full name.