SuRE-snakemake: feature: process per chromosome
Now the pipeline processes the data per chromosome after the first
alignment step. This means smaller target data sets can be generated,
and (I hope) the processing is more efficiently parallelized.

The changes are mostly in the below rules but a small number of
adaptations are made to the individual scripts as well.