code/iPCR-bam2bedpe.bash; minor change
The script now explicitly compresses the generated bedpe file as
processed in the same script, instead of compressing all files with
extension .bedpe. This was necessary since data is processed in a per
chromosome manner; which means multiple files with extension bedpe exist
and some of those might still be open for writing in other processes.