SuRE-snakemake: add rules for annotating homozygous SNPs
script 'add-homozygous-baseIndentities.bash' annotates SNPs in
SuREfragments which are not read in the forw/rev read and therefor not
assigned a base identity. If such SNPs are homozygous they can be
annotated anyway, using this script.
This change in SuRE-snakemake adds/adapts rules for incorporating this
script in the pipeline